mithril_build_script

Module fake_aggregator

Structs

FakeAggregatorData

In memory representation of a folder containing data imported using the scripts/import.sh script of the fake aggregator.

Functions

extract_item_by_epoch

Takes a map of json string indexed by hashes and re-indexes them using their epoch

extract_item_list_per_epoch

Takes a JSON string containing a list of items and extracts them into a map keyed by epoch.

generate_artifact_getter

generate_artifact_per_epoch_getter

generate_epoch_array

pub(crate) fn $array_name() -> [u64; $epoch.len]

generate_ids_array

pub(crate) fn $array_name() -> [&’a str; $ids.len]

generate_list_getter

pub(crate) fn $fun_name() -> &’static str

list_json_files_in_folder

Type Aliases

ArtifactId

FileContent